Various fixes and speedup of the engine by ~70x

+++ engine/search.hpp
@@ -8,6 +8,6 @@
 
 extern uint64_t nodes;
 
-std::pair<Move, Value> search(Board &board, int depth = -1);
+std::pair<Move, Value> search(Board &board, int64_t depth = -1);
 
 uint64_t perft(Board &board, int depth);
